requirements.txt
python-documentcloud==3.0.4
requests
eyecite